move all types to a types.tsx file
