--- my-react-app/src/data-objects/Drone.ts
@@ -1,6 +1,0 @@
-export type Drone = {
-    name: String,
-    isActive: boolean,
-    size: number,
-    id: number,
-}
