--- my-react-app/src/data-objects/FullCone.ts
@@ -1,6 +1,0 @@
-import { Product } from "./Product"
-
-export type FullCone = {
-    droneId: number,
-    components: Product[]
-}
