--- my-react-app/src/data-objects/Order.ts
@@ -1,10 +1,0 @@
-import { FullCone } from "./FullCone";
-
-export type Order = {
-    cones: FullCone[],
-    totalPrice: number,
-    employeeCut: number,
-    remainder: number,
-    timestamp: Date,
-    id: number,
-}
