--- my-react-app/src/data-objects/User.ts
@@ -1,12 +1,0 @@
-export type User = {
-    username: String,
-    userType: UserType,
-    isActive: boolean,
-    id: number
-}
-
-export enum UserType {
-    CUSTOMER = "Customer",
-    EMPLOYEE = "Employee",
-    MANAGER = "Manager"
-}
